+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+# 2.1.1-dev
+
+Release date: `202x-xx-xx`
+
+## Bug Fixes
+
+- 
+
+## Features
+
+- 
+
+## Technical Changes
+
+- 
+
 # 2.1.0
 
 Release date: `2025-03-03`